AgriGuru/requirements.txt
Flask
python-dotenv
requests
google-generativeai
folium
geopy
gunicorn